Make AsyncDispatch deliver all messages even if interrupted
This makes it keep the illusion of calling the same interface
synchronously. Instead of dropping already sent messages, it
will deliver them, but set the interrupted flag while it does so.

This means that correct behavior on interrupts now depends on the
object being dispatched to. If it never returns, the AsyncDispatch
will also never return instead of leaving a zombie thread around.